Synthetic lockup refactor following changes (#896)
* synthetic lockup refactor following changes

* Comment out things I want to change

* Comment out the test consciously being broken

+++ x/superfluid/keeper/stake.go
@@ -93,12 +93,12 @@ func (k Keeper) RefreshIntermediaryDelegationAmounts(ctx sdk.Context) {
 			Duration:      time.Hour * 24 * 14,
 		})
 
-		amt := k.GetSuperfluidOSMOTokens(ctx, acc.Denom, totalSuperfluidDelegation)
-		if amt.IsZero() {
+		amount := k.GetSuperfluidOSMOTokens(ctx, acc.Denom, totalSuperfluidDelegation)
+		if amount.IsZero() {
 			continue
 		}
 
-		coins := sdk.Coins{sdk.NewCoin(bondDenom, amt)}
+		coins := sdk.Coins{sdk.NewCoin(bondDenom, amount)}
 		err = k.bk.MintCoins(ctx, minttypes.ModuleName, coins)
 		if err != nil {
 			panic(err)
@@ -115,7 +115,7 @@ func (k Keeper) RefreshIntermediaryDelegationAmounts(ctx sdk.Context) {
 			k.Logger(ctx).Error(fmt.Sprintf("validator not found or %s", acc.ValAddr))
 			continue
 		}
-		_, err = k.sk.Delegate(cacheCtx, mAddr, amt, stakingtypes.Unbonded, validator, true)
+		_, err = k.sk.Delegate(cacheCtx, mAddr, amount, stakingtypes.Unbonded, validator, true)
 		if err != nil {
 			// this could happen when validator is fully slashed
 			k.Logger(ctx).Error(err.Error())
@@ -221,12 +221,12 @@ func (k Keeper) SuperfluidDelegate(ctx sdk.Context, sender string, lockID uint64
 
 	// mint OSMO token based on TWAP of locked denom to denom module account
 	bondDenom := k.sk.BondDenom(ctx)
-	amt := k.GetSuperfluidOSMOTokens(ctx, acc.Denom, lock.Coins.AmountOf(acc.Denom))
-	if amt.IsZero() {
+	amount := k.GetSuperfluidOSMOTokens(ctx, acc.Denom, lock.Coins.AmountOf(acc.Denom))
+	if amount.IsZero() {
 		return types.ErrOsmoEquivalentZeroNotAllowed
 	}
 
-	coins := sdk.Coins{sdk.NewCoin(bondDenom, amt)}
+	coins := sdk.Coins{sdk.NewCoin(bondDenom, amount)}
 	err = k.bk.MintCoins(ctx, minttypes.ModuleName, coins)
 	if err != nil {
 		return err
@@ -246,7 +246,7 @@ func (k Keeper) SuperfluidDelegate(ctx sdk.Context, sender string, lockID uint64
 	if !found {
 		return stakingtypes.ErrNoValidatorFound
 	}
-	_, err = k.sk.Delegate(ctx, mAddr, amt, stakingtypes.Unbonded, validator, true)
+	_, err = k.sk.Delegate(ctx, mAddr, amount, stakingtypes.Unbonded, validator, true)
 	if err != nil {
 		k.Logger(ctx).Error(err.Error())
 		return err
@@ -278,15 +278,6 @@ func (k Keeper) SuperfluidDelegate(ctx sdk.Context, sender string, lockID uint64
 
 func (k Keeper) SuperfluidUndelegate(ctx sdk.Context, sender string, lockID uint64) (sdk.ValAddress, error) {
 
-	lock, err := k.lk.GetLockByID(ctx, lockID)
-	if err != nil {
-		return nil, err
-	}
-
-	if lock.Owner != sender {
-		return nil, lockuptypes.ErrNotLockOwner
-	}
-
 	// Remove previously created synthetic lockup
 	intermediaryAccAddr := k.GetLockIdIntermediaryAccountConnection(ctx, lockID)
 	if intermediaryAccAddr.Empty() {
@@ -294,20 +285,23 @@ func (k Keeper) SuperfluidUndelegate(ctx sdk.Context, sender string, lockID uint
 	}
 	intermediaryAcc := k.GetIntermediaryAccount(ctx, intermediaryAccAddr)
 	suffix := stakingSuffix(intermediaryAcc.ValAddr)
+
+	synthLock, err := k.lk.GetSyntheticLockup(ctx, lockID, suffix)
+	if err != nil {
+		return nil, err
+	}
+
+	if synthLock.Owner != sender {
+		return nil, lockuptypes.ErrNotLockOwner
+	}
+
 	err = k.lk.DeleteSyntheticLockup(ctx, lockID, suffix)
 	if err != nil {
 		return nil, err
 	}
 
-	params := k.GetParams(ctx)
-	suffix = unstakingSuffix(intermediaryAcc.ValAddr)
-	unlocking := true
-	err = k.lk.CreateSyntheticLockup(ctx, lockID, suffix, params.UnbondingDuration, unlocking)
-	if err != nil {
-		return nil, err
-	}
-
-	amt := k.GetSuperfluidOSMOTokens(ctx, intermediaryAcc.Denom, lock.Coins.AmountOf(intermediaryAcc.Denom))
+	// use synthetic lockup coins for unbonding
+	amount := k.GetSuperfluidOSMOTokens(ctx, intermediaryAcc.Denom, synthLock.Coins.AmountOf(intermediaryAcc.Denom+suffix))
 
 	valAddr, err := sdk.ValAddressFromBech32(intermediaryAcc.ValAddr)
 	if err != nil {
@@ -315,7 +309,7 @@ func (k Keeper) SuperfluidUndelegate(ctx sdk.Context, sender string, lockID uint
 	}
 
 	shares, err := k.sk.ValidateUnbondAmount(
-		ctx, intermediaryAcc.GetAccAddress(), valAddr, amt,
+		ctx, intermediaryAcc.GetAccAddress(), valAddr, amount,
 	)
 
 	if err != nil {
@@ -327,6 +321,17 @@ func (k Keeper) SuperfluidUndelegate(ctx sdk.Context, sender string, lockID uint
 		if err != nil {
 			return valAddr, err
 		}
+	}
+
+	params := k.GetParams(ctx)
+	suffix = unstakingSuffix(intermediaryAcc.ValAddr)
+
+	// Note: bonding synthetic lockup amount is always same as native lockup amount in current implementation.
+	// If there's the case, it's different, we should create synthetic lockup at deleted bonding
+	// synthetic lockup amount
+	err = k.lk.CreateSyntheticLockup(ctx, lockID, suffix, params.UnbondingDuration, true)
+	if err != nil {
+		return nil, err
 	}
 
 	k.DeleteLockIdIntermediaryAccountConnection(ctx, lockID)
